+++ corpus/issue-1996/.clj-kondo/hooks.clj
@@ -12,13 +12,15 @@
   (when-not v
     (api/reg-finding! (assoc (meta expr)
                              :type :syntax
-                             :message "Unresolved"))))
+                             :message (str "Unresolved: " expr)))))
 
 (defn my-test [{:keys [node]}]
   (let [[core-wrd foo-wrd
          inc-node
-         dude-node] (rest (:children node))]
+         dude-node
+         local] (-> (rest (:children node)))]
     (lint-with-redefs (api/resolve {:name (api/sexpr (first (:children core-wrd)))}) core-wrd)
     (lint-with-redefs (api/resolve {:name (api/sexpr (first (:children foo-wrd)))}) foo-wrd)
     (lint-unresolved (api/resolve {:name (api/sexpr (first (:children inc-node)))}) inc-node)
-    (lint-unresolved (api/resolve {:name (api/sexpr (first (:children dude-node)))}) dude-node)))
+    (lint-unresolved (api/resolve {:name (api/sexpr (first (:children dude-node)))}) (first (:children dude-node)))
+    (lint-unresolved (api/resolve {:name (api/sexpr local)}) local)))